pyproject.toml
[project]
name = "greenwelt"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.0",
    "pygame-ce>=2.5.0",
    "pygbag>=0.9.2",
    "tomli>=2.0.1",
]

[tool.pygbag]
exclude = [
    "tools",
    "plan",
    "build",
    "docs",
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
    "build.py",
]
